move usage meta to a separate mod
